Move ProviderType enum to server-common and clarify OAuth provider config supports 'default' and 'oidc'
